Go/Grpc/python/pyproject.toml
[project]
name = "pythongrpc"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "grpcio>=1.67.0",
    "loguru>=0.7.2",
    "protobuf>=5.28.3",
]

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"
cache-dir = "./.uv_cache"
link-mode = "copy"
python-downloads = "manual"

[tool.uv.pip]
index-url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.pyright]
venvPath = "."
venv = ".venv"
#指定一个执行环境，其中src是根路径。pyright将在执行时将该路径加到sys.path。
executionEnvironments = [
  { root = "src" }
]
